Endpoint/pyproject.toml
[tool.poetry]
name = "endpoint"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
"RPi.GPIO2" = "^0.3.0-alpha.3"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
